scaffold

+++ app/pages/venues/[venueId].tsx
@@ -3,8 +3,11 @@ import { Head, Link, useRouter, useQuery, useParam, BlitzPage, useMutation, Rout
 import Layout from "app/core/layouts/Layout"
 import getVenue from "app/venues/queries/getVenue"
 import deleteVenue from "app/venues/mutations/deleteVenue"
+import { useCurrentUser } from "app/core/hooks/useCurrentUser"
+import AddTimeSlotModal from "app/core/components/Modals/AddTimeSlotModal"
 
 export const Venue = () => {
+  const currentUser = useCurrentUser()
   const router = useRouter()
   const venueId = useParam("venueId", "number")
   const [deleteVenueMutation] = useMutation(deleteVenue)
@@ -13,23 +16,42 @@ export const Venue = () => {
   return (
     <>
       <Head>
-        <title>Venue {venue.id}</title>
+        <title>{venue.name}</title>
       </Head>
 
       <div>
-        <h1>Venue {venue.id}</h1>
-        <pre>{JSON.stringify(venue, null, 2)}</pre>
+        <h1>{venue.name}</h1>
+        <ul>
+          {venue.timeSlots.map((timeslot) => (
+            <li key={timeslot.id}>
+              <>
+                {timeslot.id}:{timeslot.start} - {timeslot.end} votes
+              </>
+            </li>
+          ))}
+        </ul>
+
+        <Link href={Routes.BookVenuePage({ venueId: venue.id })}>
+          <a>Book </a>
+        </Link>
+
+        <AddTimeSlotModal />
 
         <Link href={Routes.EditVenuePage({ venueId: venue.id })}>
-          <a>Edit</a>
+          <a> Edit </a>
         </Link>
 
         <button
           type="button"
           onClick={async () => {
-            if (window.confirm("This will be deleted")) {
-              await deleteVenueMutation({ id: venue.id })
-              router.push(Routes.VenuesPage())
+            if (currentUser) {
+              if (window.confirm("This will be deleted")) {
+                await deleteVenueMutation({ id: venue.id })
+                router.push(Routes.VenuesPage())
+              }
+            } else {
+              if (window.confirm("You must be signed in to delete venues")) {
+              }
             }
           }}
           style={{ marginLeft: "0.5rem" }}
@@ -57,7 +79,7 @@ const ShowVenuePage: BlitzPage = () => {
   )
 }
 
-ShowVenuePage.authenticate = true
+ShowVenuePage.authenticate = false
 ShowVenuePage.getLayout = (page) => <Layout>{page}</Layout>
 
 export default ShowVenuePage
